locales | split sql editor into components
