+++ app/assets/stylesheets/module-data.scss
@@ -10,6 +10,7 @@
 
   &-home-nav--tab {
     list-style: none;
+
     display: inline-block;
     color: var(--color-base);
     text-transform: uppercase;
@@ -317,7 +318,7 @@
   }
 
   .codemirror {
-    border:0;
+    border: 0;
   }
 
   .CodeMirror {
@@ -336,42 +337,32 @@
 
   &-sql-editor {
     margin: 1rem 0;
+
     .btn-sql-editor {
       text-transform: none;
       min-height: 31px;
       border: 1px solid rgba(var(--color-base-string), .2);
       font-weight: 600;
       font-size: 14px;
+
       &.isDisabled {
         opacity: .4;
         pointer-events: none;
       }
+
       &.remove-label {
         font-size: 0;
+
         i {
           font-size: 14px;
         }
       }
+
       &.btn-sql-editor-run {
         float: right;
       }
     }
-  }
 
-  &-sql-editor-toolbar {
-    padding: .5rem 0;
-  }
-
-  &-sql-editor-container-save {
-    display: inline-block;
-    padding: 0 .5rem;
-  }
-
-  &-sql-editor-container-save-checkbox {
-    display: inline-block;
-  }
-
-  &-sql-editor {
     input[type="text"] {
       background: #f9f9f9;
       border: 1px solid rgba(0, 89, 201, .2);
@@ -379,7 +370,8 @@
       border-radius: 3px;
       display: inline-block;
       padding: .25rem;
-      &.query-saved {
+
+      &.disable-input-text {
         border-color: #f9f9f9;
         pointer-events: none;
       }
@@ -391,13 +383,32 @@
     }
 
     input[type="checkbox"] + label {
+      padding: 0 .25rem;
+
       &:hover {
         background-color: transparent;
       }
     }
+
     input[type="checkbox"]:checked + label {
       background-color: transparent;
     }
+  }
+
+  &-sql-editor-toolbar {
+    padding: .5rem 0;
+  }
+
+  &-sql-editor-container-save {
+    display: inline-block;
+    padding: 0 .5rem;
+    i {
+      min-width: 20px;
+    }
+  }
+
+  &-sql-editor-container-save-checkbox {
+    display: inline-block;
   }
 
   &-sql-editor-container-save-text {
